Add 64-bit support to AcpiRead and AcpiWrite.
Needed by drivers for new ACPi tables.
Internal versions of these functions still use 32-bit max transfers,
in order to minimize disruption and stack use for the standard ACPI
registers (FADT-based).

+++ source/include/acpixf.h
@@ -573,12 +573,12 @@ AcpiReset (
 
 ACPI_STATUS
 AcpiRead (
-    UINT32                  *Value,
+    UINT64                  *Value,
     ACPI_GENERIC_ADDRESS    *Reg);
 
 ACPI_STATUS
 AcpiWrite (
-    UINT32                  Value,
+    UINT64                  Value,
     ACPI_GENERIC_ADDRESS    *Reg);
 
 ACPI_STATUS
